fix(plugin code): Refactored code of Plugin
Refactored code of Plugin
